Move specs to be side-by-side with their source
- Updated how specs are run in light of this change
- Add clean npm tasks for entire `dist` directory and specs only in `dist` directory
