+++ .eslintignore
@@ -1,4 +1,4 @@
 **/node_modules
-/specs
 /examples
-/tmp
+**/*.specs.js
+/dist
