+++ src/V1Server.specs.js
@@ -2,9 +2,9 @@ import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 chai.use(chaiAsPromised);
 let should = chai.should();
-import * as Sut from './../dist/V1Server';
+import * as Sut from './V1Server';
 
-describe('src/V1Server', function () {
+describe('src/V1Server', function() {
 	let actual;
 	describe('given a V1 instance server url, protocol, port, username, password', () => {
 		let v1ServerInfo;
